Major revamps from arrays to node and edges to represent playing field. Also changed the rendering to screen to show field locations and traversal cost separately
Field was previously represented by an array but this is inflexible when playing field is no a grid. So moved to an node and edge representation. Introduced the scene which contains the field and entities, this is different from before where the field contained the entities. Created a set of primitive classes including edge, node, and removed point2d and replaced with more generic dimension2d which can be use for dimensions, weight.